example/requirements.txt
Flask==0.11.1
flask-cors==3.0.2
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.11.10
click==6.6
itsdangerous==0.24
errorgeopy==1.0.0
